+++ common/src/main/java/com/wulian/awesomesheepswell/AwesomeSheepSwell.java
@@ -6,6 +6,8 @@ import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
 import net.minecraft.entity.EntityType;
 
+import java.util.Random;
+
 public class AwesomeSheepSwell {
     public static final String MOD_ID = "awesomesheepswell";
     public static Config config;
@@ -18,5 +20,24 @@ public class AwesomeSheepSwell {
 
     public static int getMaxThickness() {
         return config.maxThickness;
+    }
+
+    public static int getRandomThickness() {
+        Random random = new Random();
+        int[] weights = {10, 9, 8, 7, 6, 5, 4, 3, 2, 1};
+        int totalWeight = 0;
+
+        for (int weight : weights) {
+            totalWeight += weight;
+        }
+
+        int rnd = random.nextInt(totalWeight);
+        for (int i = 0; i < weights.length; i++) {
+            rnd -= weights[i];
+            if (rnd < 0) {
+                return i + 1;
+            }
+        }
+        return 1;
     }
 }