+++ common/src/main/java/com/wulian/awesomesheepswell/Config.java
@@ -5,6 +5,7 @@ import me.shedaniel.autoconfig.annotation.ConfigEntry;
 
 @me.shedaniel.autoconfig.annotation.Config(name = AwesomeSheepSwell.MOD_ID)
 public class Config implements ConfigData {
-    @ConfigEntry.BoundedDiscrete(min = 0, max = 32)
+    @ConfigEntry.BoundedDiscrete(min = 5, max = 32)
+    @ConfigEntry.Gui.Tooltip
     public int maxThickness = 20;
 }