+++ common/src/main/java/com/wulian/awesomesheepswell/mixin/SheepMixin.java
@@ -2,7 +2,9 @@ package com.wulian.awesomesheepswell.mixin;
 
 import com.wulian.awesomesheepswell.AwesomeSheepSwell;
 import com.wulian.awesomesheepswell.IThickness;
+import net.minecraft.entity.EntityData;
 import net.minecraft.entity.ItemEntity;
+import net.minecraft.entity.SpawnReason;
 import net.minecraft.entity.data.DataTracker;
 import net.minecraft.entity.data.DataTracker.Builder;
 import net.minecraft.entity.data.TrackedData;
@@ -10,11 +12,14 @@ import net.minecraft.entity.data.TrackedDataHandlerRegistry;
 import net.minecraft.entity.passive.SheepEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.sound.SoundCategory;
+import net.minecraft.world.LocalDifficulty;
+import net.minecraft.world.ServerWorldAccess;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.Random;
 
@@ -69,14 +74,19 @@ public abstract class SheepMixin implements IThickness {
         }
     }
 
+    @Inject(method = "initialize", at = @At("TAIL"))
+    private void injectInitialize(ServerWorldAccess world, LocalDifficulty difficulty, SpawnReason spawnReason, EntityData entityData, CallbackInfoReturnable<EntityData> cir) {
+        setThickness(AwesomeSheepSwell.getRandomThickness());
+    }
+
     @Unique
     public int getThickness() {
         return Math.min(getSheep().getDataTracker().get(THICKNESS), AwesomeSheepSwell.getMaxThickness());
     }
 
     @Unique
-    public void setThickness(int chonkyness) {
-        chonkyness = Math.min(AwesomeSheepSwell.getMaxThickness(), chonkyness);
-        getSheep().getDataTracker().set(THICKNESS, chonkyness);
+    public void setThickness(int thickness) {
+        thickness = Math.min(AwesomeSheepSwell.getMaxThickness(), thickness);
+        getSheep().getDataTracker().set(THICKNESS, thickness);
     }
 }