+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# Changelogs
+# Changelog
 
-## 1.0.0
+## 1.1.0
 
-- Initial release
+- Fixed incorrect configuration descriptions.
+- Sheep now spawn with a random thickness from 1 to 10, with decreasing weight as thickness increases.
+  Fixed a bug where baby sheep could be sheared in the Forge version.